gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "httpx>=0.28",
    "python-jose[cryptography]>=3.5.0",
    "pydantic-settings>=2.12.0",
]